requirements.txt
django==1.6.1
south
django-tagging
django-mobile
django-reversion
django-bootstrap3
